Add custom field caching to several functions.
Querying the custom fields can be a expensive task, particularly when run
hundreds and/or thousands of times. Cache the results intelligently for
performance gains in the following places:

* _CRM_Import_Parser_Contact::formatCommonData
* api/v2/utils.php::_civicrm_add_formatted_location_blocks

This wouldn't be needed but the CRM_Core_BAO_CustomField::getFields() cache
key takes quite a while to build (because permissions aren't cached). Simple
fix, but I'm not sure if permissions should be cached. This is easier, atleast
for now.