Construct the dedupe query dynamically.
This makes the process more flexible. If a Carmen R Gonzalez is already in the
system and we do an import with for Carmen Gonzalez at the same zip/address we
can match her now. Likewise, if for some reason the reverse happened, we'd
still get a match where we'd check for a null middle name instead previously.