Fixes major bug introduced into the Level 1 import
When reorganizing the dedupe code a couple commits ago I also reverted the
Level 1 dedupe back to being less agressive as discussed by Brian, Bill and
I in issue #4053. In the process I erroneously used = NULL instead of IS NULL
which causes a general failure to find duplicates.

This bug did not destroy any data, but significacntly reduce the effectiveness
of the native importer.