Adds a simple REST API for tag operations.
Currently implements create, update, and delete at the following paths:

    civicrm/ajax/tag/create
    civicrm/ajax/tag/update
    civicrm/ajax/tag/delete

These operations may receive, as $_GET parameters, any number of parameters
corresponding to field entries in the civicrm_tag table and will apply
them to the operation intelligently. Update and Delete require the id=tag_id
argument to be used. Create (obviously) forbids the tag argument.

On success, a response with code == 1 will be returned with the $tag->id of
the tag that was operated on (the new id for the create operation) in the
message field.

On failure, the response code is != 1 and the message describes the database
level problem as provided by the DAO layer.

Currently, errors are not handled very well and requirements are not checked
before performing operations. These aspects must be cleaned up later...