Excessive spacing removed from generated report HTML
The constituent summary and detail reports generate HTML to pass through to the wkhtmltopdf converter. Smarty's code was generating excessive amounts of spacing between HTML tags and in the data itself. When CiviCRM has Debug enabled, the generated HTML will be as it was originally. When Debug is disabled, it will generate smaller HTML files which should allow for much more data to appear in reports. The Smarty template cache should be cleared before testing this.