Remove excess whitespace from Smarty templates
Excess whitespace is removed from HTML generated by Smarty templates by adding {strip} tags to every template with a prefilter. Changes also needed to be made to the jquery.files.tpl and jquery loader to separate filenames on a semi-colon instead of space since the template is now being stripped of spaces. This was causing the JavaScript to not load and menus to fail. Tested script after a menu rebuild and worked as normal.