Fixes JavaScript error on non-Contact Summary pages
Non-Contact Summary pages were getting a JavaScript error because there was a line that was not not getting populated, resulting in JS code like:

var contactId = ;

which is invalid JavaScript. This happened when a contactId variable in PHP was not set due to not being on a contact page. The template now checks if contactId is populated, if the variable is populated, the template will put in the AJAX call to get the change log commit. If the variable is not populated, it will not write the AJAX call to the page.